optimrust v2.0 greedy and simplex

+++ src/node.rs
@@ -2,11 +2,11 @@
 #[derive(Debug, PartialEq, Copy, Clone)]
 pub struct Node {
     pub s_node_id: usize,
-    pub s_qty: u16,
+    pub s_qty: i32,
     pub d_node_id: usize,
-    pub d_qty: u16,
-    pub node_cost: f32,
-    pub node_qty: u16,
+    pub d_qty: i32,
+    pub node_cost: i32,
+    pub node_qty: i32,
 }
 
 impl Node {
@@ -16,12 +16,12 @@ impl Node {
             s_qty: 0,
             d_node_id: 0,
             d_qty: 0,
-            node_cost: 0.0,
+            node_cost: 0,
             node_qty: 0,
         }
     }
 
-    pub fn new_with_data(s_id: &str, d_id: &str, s_qt: u16, d_qt: u16, cost: &str) -> Node{
+    pub fn new_with_data(s_id: &str, d_id: &str, s_qt: i32, d_qt: i32, cost: &str) -> Node{
         Node {
             s_node_id: s_id.parse().unwrap(),
             s_qty: s_qt,
